Changed Forms and Classrooms
Forms is now more straightforward, redundant information is gone and the information is clearer. For classes, each class has more information pointing to where the newsletter is that contains the information of past classes. Curriculum overview was also altered for this. Teddy bears was in particular changed due to confusing out of order information. Header priorities were changed to look better. Schedule and classroom were changed, with the latter now linking to each class webpage and showing when it was.